Move client auto update helper to lib package
